Implement new bot architecture with modular components and Cypress test support
