+++ src/starbunk/bots/reply-bots/niceBot/niceBot.ts
@@ -1,21 +1,32 @@
 import webhookService, { WebhookService } from '@/webhooks/webhookService';
 import { BotBuilder } from '../../botBuilder';
 import ReplyBot from '../../replyBot';
-import { PatternCondition } from '../../triggers/conditions/patternCondition';
-import { Patterns } from '../../triggers/conditions/patterns';
-import { NICE_BOT_AVATAR_URL, NICE_BOT_RESPONSE } from './niceBotModel';
+import { BOT_NAME, NICE_BOT_AVATAR_URL } from './niceBotModel';
 
+// Import the extracted components
+import { SixtyNineCondition } from './conditions/sixtyNineCondition';
+import { NiceResponseGenerator } from './responses/niceResponseGenerator';
+
+/**
+ * Creates a NiceBot instance that responds with "Nice." to messages containing "69"
+ *
+ * @param webhookSvc - The webhook service to use for sending messages
+ * @returns A configured ReplyBot instance
+ */
 export default function createNiceBot(
 	webhookSvc: WebhookService = webhookService
 ): ReplyBot {
-	// Always use the imported singleton webhookService, ignoring any webhookService in config
-	// This ensures we're using the properly initialized webhookService with the writeMessage method
-	return new BotBuilder('NiceBot', webhookSvc)
+	// Create the condition and response generator
+	const sixtyNineCondition = new SixtyNineCondition();
+	const niceResponseGenerator = new NiceResponseGenerator();
+
+	// Build and return the bot
+	return new BotBuilder(BOT_NAME, webhookSvc)
 		.withAvatar(NICE_BOT_AVATAR_URL)
-		.withCustomCondition(
-			NICE_BOT_RESPONSE,
+		.withConditionResponse(
+			niceResponseGenerator,
 			NICE_BOT_AVATAR_URL,
-			new PatternCondition(Patterns.SPECIAL_NICE_NUMBER)
+			sixtyNineCondition
 		)
 		.build();
 }
