+++ src/starbunk/bots/reply-bots/niceBot/responses/niceResponseGenerator.ts
@@ -1,0 +1,11 @@
+import { ResponseGenerator } from "@/starbunk/bots/botTypes";
+import { NICE_BOT_RESPONSE } from "../niceBotModel";
+
+/**
+ * Response generator that provides the "Nice." response
+ */
+export class NiceResponseGenerator implements ResponseGenerator {
+	async generateResponse(): Promise<string> {
+		return NICE_BOT_RESPONSE;
+	}
+}
